+++ src/GameEngine/GameState/Player.hpp
@@ -14,12 +14,12 @@ class Command; //forward declaration
 struct Player
 {
     GameState* state;
-    unsigned id;
-    int command_score;
-    int health;
-    int currentWormId; //1-indexed
-    int consecutiveDoNothingCount;
-    int remainingWormSelections;
+    uint8_t id;
+    int16_t command_score;
+    int16_t health;
+    int8_t currentWormId; //1-indexed
+    int8_t consecutiveDoNothingCount;
+    int8_t remainingWormSelections;
     std::array<Worm, 3> worms;
     std::shared_ptr<Command> previousCommand;
